SentientsSatsLocal/requirements-extra.txt
# CORE FRAMEWORK
uagents-ai-engine==0.4.0

# API & HTTP REQUESTS
aiohttp==3.9.1
httpx==0.25.2

# DATA PARSING & PROCESSING
beautifulsoup4==4.12.2
lxml==4.9.3



# NLP & SENTIMENT ANALYSIS
nltk==3.8.1
vaderSentiment==3.3.2

# DATE & TIME HANDLING
python-dateutil==2.8.2
pytz==2023.3.post1

# DATA STRUCTURES & CACHING
diskcache==5.6.3

# ASYNC & CONCURRENCY
aiofiles==23.2.1

# LOGGING & MONITORING
loguru==0.7.2

# WEB3 & BLOCKCHAIN
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# DATA ANALYSIS & PROCESSING
pandas==2.1.4
numpy==1.26.2

# JSON & DATA SERIALIZATION
orjson==3.9.10
ujson==5.9.0

# RATE LIMITING & THROTTLING
ratelimit==2.2.1
backoff==2.2.1

# TESTING
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# CODE QUALITY
black==23.12.1
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# UTILITIES
click==8.1.7
rich==13.7.0
tabulate==0.9.0

# SECURITY
cryptography==41.0.7

# COMPATIBILITY & DEPENDENCIES
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0
